requirements.txt
PyQt5==5.15.4
Pillow==9.1.1
opencv-python==4.5.3.56
numpy==1.21.0
